1.1.0

+++ src/API.hpp
@@ -105,7 +105,7 @@ private:
                     if (isGzip) {
                         // gzip解压缩
                         uint8_t* outBuf = NULL;
-                        size_t outLen = 0;
+                        uint32_t outLen = 0;
                         ArduinoUZlib::decompress((uint8_t*)s.c_str(), (uint32_t)s.length(), outBuf, outLen);
                         error = deserializeJson(doc, (char*)outBuf, outLen);
                         free(outBuf);
